safe: move class to responses folder
It was the only class at the folder, the organization made no sense
