+++ site/BusinessLogic/Response/ChangePasswordInfo.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using DFM.BusinessLogic.InterfacesAndBases;
 
 namespace DFM.BusinessLogic.Response
 {
