+++ site/BusinessLogic/Response/SignUpInfo.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using DFM.BusinessLogic.InterfacesAndBases;
 using DFM.Entities;
 
 namespace DFM.BusinessLogic.Response
