+++ site/BusinessLogic/Services/SafeService.cs
@@ -7,7 +7,6 @@ using DFM.BusinessLogic.Repositories;
 using DFM.Entities;
 using DFM.Entities.Enums;
 using DFM.Authentication;
-using DFM.BusinessLogic.InterfacesAndBases;
 using DFM.BusinessLogic.Response;
 using DFM.Entities.Bases;
 using Keon.Util.Extensions;
